backend/requirements.txt
# FastAPI and dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0
alembic==1.12.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt>=4.0.0,<5.0.0
python-dotenv==1.0.0

# HTTP Client
httpx==0.25.1

# Azure AD OAuth2
msal==1.25.0

# Configuration
pydantic-settings==2.1.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.1  # For testing HTTP requests

# Code Quality
ruff==0.1.6
mypy==1.7.1
types-passlib==1.7.7.13
